Found a way to become member or admin by checking if the id exists between users and messages, otherwise you can still be member or admin if the id is equal to the parameter

+++ db/messageQueries.js
@@ -16,11 +16,24 @@ async function postMessageCreate(title, content, user_id) {
 
 async function postMessageDelete(id) {
   try {
-    await pool.query(
-      // "DELETE FROM messages WHERE id IN (SELECT id FROM users WHERE $1 = id AND membership_status = 'admin')",
-      "DELETE FROM messages WHERE id = $1",
-      [id]
+    const { rows } = await pool.query(
+      "SELECT user_id FROM messages WHERE user_id IN (SELECT id FROM users WHERE id = user_id)"
     );
+
+    if (rows[0]) {
+      const query = await pool.query(
+        "DELETE FROM messages WHERE user_id IN (SELECT users.id FROM users WHERE user_id = $1 OR membership_status = 'admin')",
+        [id]
+      );
+
+      return query.rows[0];
+    } else {
+      const query = await pool.query("DELETE FROM messages WHERE id = $1", [
+        id,
+      ]);
+
+      return query.rows[0];
+    }
   } catch (err) {
     console.error("Error deleting a message, user is not an admin", err);
     throw err;
